Draft review: Add note to List of Tables about supplement.
5. We recommend that you add a note to the page with the “List of
Tables” informing the reader that, in many cases, the captions for
tables are provided in the contents of the dissertation but the actual
tables are available in a supplementary file accompanying the
dissertation (e.g., “Note: Due to the large size of some tables, the
tables have been provided in a supplementary file accompanying the
dissertation. In such cases, the page number provided below directs the
reader to a table’s caption.”).