Draft review: Change extension of tds files to txt
i. We were unable to open the TDS Files in the supplementary file. You
may include this file type, but please be aware that the material may be
inaccessible to some users. If possible, and not prohibitively
time-consuming, we recommend that you provide these tables in an
alternate format, such as Excel or Text Document, in addition to TDS.